--- connect-webapp/src/app/components/register-user/register-user.page.html
@@ -1,55 +1,0 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <ion-title>register-user</ion-title>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content [fullscreen]="true">
-  <form [formGroup]="userRegistrationForm" (ngSubmit)="register()">
-
-    <ion-item>
-      <ion-label position="floating">First Name</ion-label>
-      <ion-input type="text" formControlName="firstName"></ion-input>
-      <ion-text color="danger" *ngIf="firstName?.invalid && (firstName?.dirty || firstName?.touched)">First name is required.</ion-text>
-    </ion-item>
-
-    <ion-item>
-      <ion-label position="floating">Last Name</ion-label>
-      <ion-input type="text" formControlName="lastName"></ion-input>
-      <ion-text color="danger" *ngIf="lastName?.invalid && (lastName?.dirty || lastName?.touched)">Last name is required.</ion-text>
-    </ion-item>
-
-
-    <ion-item>
-      <ion-label position="floating">Email</ion-label>
-      <ion-input type="email" formControlName="email"></ion-input>
-    </ion-item>
-
-    <ion-item>
-      <ion-label position="floating">Password</ion-label>
-      <ion-input type="password" formControlName="password"></ion-input>
-    </ion-item>
-
-    <ion-item *ngIf="email?.invalid && (email?.dirty || email?.touched)">
-      <ion-text color="danger" *ngIf="email?.errors?. ['required']">Email is required.</ion-text>
-      <ion-text color="danger" *ngIf="email?.errors?.['email']">Please enter a valid email address.</ion-text>
-    </ion-item>
-
-    <ion-item>
-      <ion-label position="floating">Phone Number</ion-label>
-      <ion-input type="text" formControlName="phoneNumber"></ion-input>
-      <ion-text color="danger" *ngIf="phoneNumber?.invalid && (phoneNumber?.dirty || phoneNumber?.touched)">Phone number is required.</ion-text>
-    </ion-item>
-
-    <ion-item>
-      <ion-label position="floating">Bio (Max 50 characters)</ion-label>
-      <ion-textarea formControlName="bio"></ion-textarea>
-      <ion-text color="danger" *ngIf="bio?.invalid && (bio?.dirty || bio?.touched)">
-        <div *ngIf="bio?.errors?.['maxLength']">Bio cannot exceed 50 characters.</div>
-      </ion-text>
-    </ion-item>
-
-
-    <ion-button expand="full" type="submit" >Register</ion-button>
-  </form>
-</ion-content>
