--- connect-webapp/src/app/components/register-user/register-user.page.spec.ts
@@ -1,17 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { RegisterUserPage } from './register-user.page';
-
-describe('RegisterUserPage', () => {
-  let component: RegisterUserPage;
-  let fixture: ComponentFixture<RegisterUserPage>;
-
-  beforeEach(async(() => {
-    fixture = TestBed.createComponent(RegisterUserPage);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  }));
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
