+++ connect-webapp/src/app/components/signup-client/signup-client.page.scss
@@ -1,0 +1,16 @@
+.container {
+  display: flex;
+  justify-content: center;
+  align-items: start;
+  padding-top: 100px;
+  height: 100%;
+}
+
+ion-card{
+  min-width: 50% !important;
+  margin: 0 !important;
+}
+
+.back{
+  background: radial-gradient(circle at 18.7% 37.8%, rgb(250, 250, 250) 0%, rgb(225, 234, 238) 90%) !important;
+}
