+++ connect-webapp/src/theme/variables.scss
@@ -153,6 +153,7 @@
    */
 
   .ios body {
+    --ion-background: radial-gradient(circle at 18.7% 37.8%, rgb(250, 250, 250) 0%, rgb(225, 234, 238) 90%);
     --ion-background-color: #000000;
     --ion-background-color-rgb: 0,0,0;
 
@@ -197,6 +198,7 @@
    */
 
   .md body {
+    --ion-background: radial-gradient(circle at 18.7% 37.8%, rgb(250, 250, 250) 0%, rgb(225, 234, 238) 90%);
     --ion-background-color: #121212;
     --ion-background-color-rgb: 18,18,18;
